Set imported transient workflow progress and initialization state
